fix bug with multiple persons
