+++ posthog/models/filters/stickiness_filter.py
@@ -4,13 +4,15 @@ from django.db.models.functions.datetime import TruncDay, TruncHour, TruncMinute
 from django.http import HttpRequest
 
 from posthog.models.filters.base_filter import BaseFilter
-from posthog.models.filters.mixins.common import CompareMixin, InsightMixin, OffsetMixin, ShownAsMixin
-from posthog.models.filters.mixins.property import PropertyMixin
-from posthog.models.filters.mixins.stickiness import (
-    SelectedIntervalMixin,
+from posthog.models.filters.mixins.common import (
+    CompareMixin,
+    InsightMixin,
+    OffsetMixin,
+    ShownAsMixin,
     TargetEntityDerivedMixin,
-    TotalIntervalsDerivedMixin,
 )
+from posthog.models.filters.mixins.property import PropertyMixin
+from posthog.models.filters.mixins.stickiness import SelectedIntervalMixin, TotalIntervalsDerivedMixin
 from posthog.models.team import Team
